fixup! fixup! Support Reschedule mode for `BaseSensorOperator` with Task SDK
