+++ airflow-core/tests/unit/api_fastapi/execution_api/routes/test_task_instances.py
@@ -1160,3 +1160,62 @@ class TestPreviousDagRun:
             "start_date": None,
             "end_date": None,
         }
+
+
+class TestGetRescheduleStartDate:
+    def test_get_start_date(self, client, session, create_task_instance):
+        ti = create_task_instance(
+            task_id="test_ti_update_state_reschedule_mysql_limit",
+            state=State.RUNNING,
+            start_date=timezone.datetime(2024, 1, 1),
+            session=session,
+        )
+        tr = TaskReschedule(
+            task_instance_id=ti.id,
+            try_number=1,
+            start_date=timezone.datetime(2024, 1, 1),
+            end_date=timezone.datetime(2024, 1, 1, 1),
+            reschedule_date=timezone.datetime(2024, 1, 1, 2),
+        )
+        session.add(tr)
+        session.commit()
+
+        response = client.get(f"/execution/task-reschedules/{ti.id}/start_date")
+        assert response.status_code == 200
+        assert response.json() == "2024-01-01T00:00:00Z"
+
+    def test_get_start_date_not_found(self, client):
+        ti_id = "0182e924-0f1e-77e6-ab50-e977118bc139"
+        response = client.get(f"/execution/task-reschedules/{ti_id}/start_date")
+        assert response.status_code == 404
+
+    def test_get_start_date_with_try_number(self, client, session, create_task_instance):
+        # Create multiple reschedules
+        dates = [
+            timezone.datetime(2024, 1, 1),
+            timezone.datetime(2024, 1, 2),
+            timezone.datetime(2024, 1, 3),
+        ]
+
+        ti = create_task_instance(
+            task_id="test_get_start_date_with_try_number",
+            state=State.RUNNING,
+            start_date=timezone.datetime(2024, 1, 1),
+            session=session,
+        )
+
+        for i, date in enumerate(dates, 1):
+            tr = TaskReschedule(
+                task_instance_id=ti.id,
+                try_number=i,
+                start_date=date,
+                end_date=date.replace(hour=1),
+                reschedule_date=date.replace(hour=2),
+            )
+            session.add(tr)
+        session.commit()
+
+        # Test getting start date for try_number 2
+        response = client.get(f"/execution/task-reschedules/{ti.id}/start_date?try_number=2")
+        assert response.status_code == 200
+        assert response.json() == "2024-01-02T00:00:00Z"
