+++ task-sdk/src/airflow/sdk/execution_time/supervisor.py
@@ -964,7 +964,7 @@ class ActivitySubprocess(WatchedSubprocess):
             dr_resp = self.client.dag_runs.get_state(msg.dag_id, msg.run_id)
             resp = DagRunStateResult.from_api_response(dr_resp).model_dump_json().encode()
         elif isinstance(msg, GetTaskRescheduleStartDate):
-            tr_resp = self.client.task_reschedules.get_start_date(msg.ti_id, msg.try_number)
+            tr_resp = self.client.task_instances.get_reschedule_start_date(msg.ti_id, msg.try_number)
             resp = tr_resp.model_dump_json().encode()
         else:
             log.error("Unhandled request", msg=msg)
