+++ task-sdk/tests/task_sdk/api/test_client.py
@@ -837,7 +837,7 @@ class TestTaskRescheduleOperations:
             return httpx.Response(status_code=422)
 
         client = make_client(transport=httpx.MockTransport(handle_request))
-        result = client.task_reschedules.get_start_date(id=ti_id, try_number=1)
+        result = client.task_instances.get_reschedule_start_date(id=ti_id, try_number=1)
 
         assert isinstance(result, TaskRescheduleStartDate)
         assert result.start_date == "2024-01-01T00:00:00Z"
